+++ modules/physics/bullet/src/bullet.cpp
@@ -27,8 +27,8 @@ const char *Bullet::version() const {
     return "1.0";
 }
 
-uint8_t Bullet::types() const {
-    uint8_t result  = SYSTEM;
+uint32_t Bullet::types() const {
+    uint32_t result  = SYSTEM;
 #ifdef NEXT_SHARED
     result  |= CONVERTER;
 #endif
