+++ modules/renders/rendergl/includes/rendergl.h
@@ -13,7 +13,7 @@ public:
 
     const char *version() const;
 
-    uint8_t types() const;
+    uint32_t types() const;
 
     System *system();
